Refactored the UI on Community Forum and Reward Page and added it to the navbar
