--- src/pages/forms/forms.css
@@ -1,11 +1,0 @@
-.app {
-    background-color: #a6dbf7;
-    height: 100vh;
-    width: 100vw;
-    overflow: auto;
-}
-
-.form-search {
-    width: 50vw;
-    margin-left: 11vh;
-}
